+++ lib/lhm.rb
@@ -58,15 +58,15 @@ module Lhm
   # @option options [Time] :until
   #   Filter to only remove tables up to specified time (defaults to: nil)
   def cleanup(run = false, options = {})
-    lhm_tables = connection.select_values("show tables").select { |name| name =~ /^lhm(a|n)_/ }
+    lhm_tables = connection.select_values('show tables').select { |name| name =~ /^lhm(a|n)_/ }
     if options[:until]
-      lhm_tables.select!{ |table|
-        table_date_time = Time.strptime(table, "lhma_%Y_%m_%d_%H_%M_%S")
+      lhm_tables.select! { |table|
+        table_date_time = Time.strptime(table, 'lhma_%Y_%m_%d_%H_%M_%S')
         table_date_time <= options[:until]
       }
     end
 
-    lhm_triggers = connection.select_values("show triggers").collect do |trigger|
+    lhm_triggers = connection.select_values('show triggers').collect do |trigger|
       trigger.respond_to?(:trigger) ? trigger.trigger : trigger
     end.select { |name| name =~ /^lhmt/ }
 
@@ -79,12 +79,12 @@ module Lhm
       end
       true
     elsif lhm_tables.empty? && lhm_triggers.empty?
-      puts "Everything is clean. Nothing to do."
+      puts 'Everything is clean. Nothing to do.'
       true
     else
-      puts "Existing LHM backup tables: #{lhm_tables.join(", ")}."
-      puts "Existing LHM triggers: #{lhm_triggers.join(", ")}."
-      puts "Run Lhm.cleanup(true) to drop them all."
+      puts "Existing LHM backup tables: #{lhm_tables.join(', ')}."
+      puts "Existing LHM triggers: #{lhm_triggers.join(', ')}."
+      puts 'Run Lhm.cleanup(true) to drop them all.'
       false
     end
   end
@@ -120,5 +120,4 @@ module Lhm
   def connection
     Connection.new(adapter)
   end
-
 end