+++ lib/lhm/chunker.rb
@@ -50,8 +50,8 @@ module Lhm
     end
 
     def copy(lowest, highest)
-      "insert ignore into `#{ destination_name }` (#{ destination_columns }) " +
-      "select #{ origin_columns } from `#{ origin_name }` " +
+      "insert ignore into `#{ destination_name }` (#{ destination_columns }) " \
+      "select #{ origin_columns } from `#{ origin_name }` " \
       "#{ conditions } `#{ origin_name }`.`id` between #{ lowest } and #{ highest }"
     end
 
@@ -65,22 +65,22 @@ module Lhm
       limit ? limit.to_i : nil
     end
 
-    #XXX this is extremely brittle and doesn't work when filter contains more
-    #than one SQL clause, e.g. "where ... group by foo". Before making any
-    #more changes here, please consider either:
+    # XXX this is extremely brittle and doesn't work when filter contains more
+    # than one SQL clause, e.g. "where ... group by foo". Before making any
+    # more changes here, please consider either:
     #
-    #1. Letting users only specify part of defined clauses (i.e. don't allow
-    #`filter` on Migrator to accept both WHERE and INNER JOIN
-    #2. Changing query building so that it uses structured data rather than
-    #strings until the last possible moment.
+    # 1. Letting users only specify part of defined clauses (i.e. don't allow
+    # `filter` on Migrator to accept both WHERE and INNER JOIN
+    # 2. Changing query building so that it uses structured data rather than
+    # strings until the last possible moment.
     def conditions
       if @migration.conditions
         @migration.conditions.
-          sub(/\)\Z/, "").
-          #put any where conditions in parens
-          sub(/where\s(\w.*)\Z/, "where (\\1)") + " and"
+          sub(/\)\Z/, '').
+          # put any where conditions in parens
+          sub(/where\s(\w.*)\Z/, 'where (\\1)') + ' and'
       else
-        "where"
+        'where'
       end
     end
 
@@ -102,7 +102,7 @@ module Lhm
 
     def validate
       if @start && @limit && @start > @limit
-        error("impossible chunk options (limit must be greater than start)")
+        error('impossible chunk options (limit must be greater than start)')
       end
     end
   end