+++ spec/integration/cleanup_spec.rb
@@ -3,11 +3,11 @@
 
 require File.expand_path(File.dirname(__FILE__)) + '/integration_helper'
 
-describe Lhm, "cleanup" do
+describe Lhm, 'cleanup' do
   include IntegrationHelper
   before(:each) { connect_master! }
 
-  describe "changes" do
+  describe 'changes' do
     before(:each) do
       table_create(:users)
       simulate_failed_migration do
@@ -22,49 +22,49 @@ describe Lhm, "cleanup" do
       Lhm.cleanup(true)
     end
 
-    it "should show temporary tables" do
+    it 'should show temporary tables' do
       output = capture_stdout do
         Lhm.cleanup
       end
-      output.must_include("Existing LHM backup tables")
+      output.must_include('Existing LHM backup tables')
       output.must_match(/lhma_[0-9_]*_users/)
     end
 
-    it "should show temporary tables within range" do
+    it 'should show temporary tables within range' do
       table = OpenStruct.new(:name => 'users')
       table_name = Lhm::Migration.new(table, nil, nil, {}, Time.now - 172800).archive_name
       table_rename(:users, table_name)
 
       output = capture_stdout do
-        Lhm.cleanup false, {:until => Time.now - 86400}
+        Lhm.cleanup false, { :until => Time.now - 86400 }
       end
-      output.must_include("Existing LHM backup tables")
+      output.must_include('Existing LHM backup tables')
       output.must_match(/lhma_[0-9_]*_users/)
     end
 
-    it "should exclude temporary tables outside range" do
+    it 'should exclude temporary tables outside range' do
       table = OpenStruct.new(:name => 'users')
       table_name = Lhm::Migration.new(table, nil, nil, {}, Time.now).archive_name
       table_rename(:users, table_name)
 
       output = capture_stdout do
-        Lhm.cleanup false, {:until => Time.now - 172800}
+        Lhm.cleanup false, { :until => Time.now - 172800 }
       end
-      output.must_include("Existing LHM backup tables")
+      output.must_include('Existing LHM backup tables')
       output.wont_match(/lhma_[0-9_]*_users/)
     end
 
-    it "should show temporary triggers" do
+    it 'should show temporary triggers' do
       output = capture_stdout do
         Lhm.cleanup
       end
-      output.must_include("Existing LHM triggers")
-      output.must_include("lhmt_ins_users")
-      output.must_include("lhmt_del")
-      output.must_include("lhmt_upd_users")
+      output.must_include('Existing LHM triggers')
+      output.must_include('lhmt_ins_users')
+      output.must_include('lhmt_del')
+      output.must_include('lhmt_upd_users')
     end
 
-    it "should delete temporary tables" do
+    it 'should delete temporary tables' do
       Lhm.cleanup(true).must_equal(true)
       Lhm.cleanup.must_equal(true)
     end