+++ spec/integration/locked_switcher_spec.rb
@@ -12,30 +12,30 @@ describe Lhm::LockedSwitcher do
 
   before(:each) { connect_master! }
 
-  describe "switching" do
+  describe 'switching' do
     before(:each) do
-      @origin = table_create("origin")
-      @destination = table_create("destination")
+      @origin = table_create('origin')
+      @destination = table_create('destination')
       @migration = Lhm::Migration.new(@origin, @destination)
     end
 
-    it "rename origin to archive" do
+    it 'rename origin to archive' do
       switcher = Lhm::LockedSwitcher.new(@migration, connection)
       switcher.run
 
       slave do
         table_exists?(@origin).must_equal true
-        table_read(@migration.archive_name).columns.keys.must_include "origin"
+        table_read(@migration.archive_name).columns.keys.must_include 'origin'
       end
     end
 
-    it "rename destination to origin" do
+    it 'rename destination to origin' do
       switcher = Lhm::LockedSwitcher.new(@migration, connection)
       switcher.run
 
       slave do
         table_exists?(@destination).must_equal false
-        table_read(@origin.name).columns.keys.must_include "destination"
+        table_read(@origin.name).columns.keys.must_include 'destination'
       end
     end
   end